+++ src/components/SSOLanding/useSSOSession.js
@@ -4,12 +4,11 @@ import { useLocation } from 'react-router-dom';
 import { useCookies } from 'react-cookie';
 import queryString from 'query-string';
 
-import { config, okapi } from 'stripes-config';
+import { okapi } from 'stripes-config';
 
 import { defaultErrors } from '../../constants';
 import { setAuthError } from '../../okapiActions';
 import { requestUserWithPerms } from '../../loginServices';
-import { parseJWT } from '../../helpers';
 
 const getParams = (location) => {
   const search = location.search;
@@ -24,9 +23,7 @@ const getToken = (cookies, params) => {
 };
 
 const getTenant = (params, token, store) => {
-  const tenant = config.useSecureTokens
-    ? params?.tenantId
-    : parseJWT(token)?.tenant;
+  const tenant = params?.tenantId;
 
   return tenant || store.getState()?.okapi?.tenant;
 };