+++ src/components/SSOLanding/useSSOSession.test.js
@@ -91,6 +91,7 @@ describe('SSOLanding', () => {
 
   it('should request user session when RTR is disabled and right tenant from ssoToken', () => {
     const tokenTenant = 'tokenTenant';
+    const okapiTenant = 'okapiTenant';
     const store = useStore();
 
     useLocation.mockReturnValue({ search: `ssoToken=${ssoTokenValue}` });
@@ -99,7 +100,7 @@ describe('SSOLanding', () => {
 
     renderHook(() => useSSOSession());
 
-    expect(requestUserWithPerms).toHaveBeenCalledWith(okapi.url, store, tokenTenant, ssoTokenValue);
+    expect(requestUserWithPerms).toHaveBeenCalledWith(okapi.url, store, okapiTenant, ssoTokenValue);
   });
 
   it('should request user session when RTR is enabled and right tenant from query params', () => {