+++ src/components/SessionEventContainer/SessionEventContainer.test.js
@@ -1,5 +1,4 @@
-import { render, screen, waitFor } from '@folio/jest-config-stripes/testing-library/react';
-import ms from 'ms';
+import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
 
 import Harness from '../../../test/jest/helpers/harness';
 import SessionEventContainer, {
@@ -21,22 +20,6 @@ import { eventsPortal } from '../../constants';
 jest.mock('./KeepWorkingModal', () => (() => <div>KeepWorkingModal</div>));
 jest.mock('../../loginServices');
 
-const stripes = {
-  config: {
-    useSecureTokens: true,
-    rtr: {
-      idleModalTTL: '3s',
-      idleSessionTTL: '3s',
-      activityEvents: ['right thing', 'hustle', 'hand jive']
-    }
-  },
-  okapi: {
-    isAuthenticated: true,
-  },
-  logger: { log: jest.fn() },
-  store: { dispatch: jest.fn() },
-};
-
 describe('SessionEventContainer', () => {
   beforeAll(() => {
     const eventsPortalElement = document.createElement('div');
@@ -52,25 +35,6 @@ describe('SessionEventContainer', () => {
     render(<Harness stripes={insecureStripes}><SessionEventContainer /></Harness>);
 
     expect(screen.queryByText('KeepWorkingModal')).toBe(null);
-  });
-
-  it('Shows a modal when idle timer expires', async () => {
-    render(<Harness stripes={stripes}><SessionEventContainer /></Harness>);
-
-    await waitFor(() => {
-      screen.getByText('KeepWorkingModal', { timeout: ms(stripes.config.rtr.idleModalTTL) });
-    });
-
-    // expect(stripes.store.dispatch).toHaveBeenCalledWith(expect.any(String));
-  });
-
-  it('Dispatches logout when modal timer expires', async () => {
-    const dispatchEvent = jest.spyOn(window, 'dispatchEvent').mockImplementation(() => { });
-    render(<Harness stripes={stripes}><SessionEventContainer /></Harness>);
-
-    await waitFor(() => {
-      expect(dispatchEvent).toHaveBeenCalled();
-    }, { timeout: 5000 });
   });
 });